refactor: put back the helper file that was taken out of utils directory by accident
