+++ src/components/layouts/GlobalLayout.tsx
@@ -5,7 +5,7 @@ import { Navigate, Outlet, useLocation } from "react-router-dom"
 import { useCookies } from "react-cookie"
 import { useGetProfileQuery } from "../../features/user/backendApi"
 import { login } from "../../features/user/userSlice"
-import { getErrorInfo, getRoleBasedHomepageURL } from "../../helper"
+import { getErrorInfo, getRoleBasedHomepageURL } from "../../utils/helper"
 import { handleShowAlert } from "../../utils/handleShowAlert"
 import Loader from "../ui/Loader"
 import { AlertType, Cookie } from "../../utils/types"
