+++ src/components/modals/AddCoach.tsx
@@ -3,7 +3,7 @@ import Button from "../ui/Button"
 import { Modal } from "@mui/material"
 import { AlertType, Cookie, User, UserRole } from "../../utils/types"
 import Select from "../ui/Select"
-import { getErrorInfo } from "../../helper"
+import { getErrorInfo } from "../../utils/helper"
 import { handleShowAlert } from "../../utils/handleShowAlert"
 import { useDispatch } from "react-redux"
 import { useCookies } from "react-cookie"
