+++ src/components/modals/CreateUser.tsx
@@ -6,7 +6,7 @@ import Input from "../ui/Input"
 import Select from "../ui/Select"
 import { z } from "zod"
 import { zodResolver } from "@hookform/resolvers/zod"
-import { getErrorInfo } from "../../helper"
+import { getErrorInfo } from "../../utils/helper"
 import { handleShowAlert } from "../../utils/handleShowAlert"
 import { useDispatch } from "react-redux"
 import { useCookies } from "react-cookie"
