+++ src/components/modals/EditParticipantModal.tsx
@@ -5,7 +5,7 @@ import { useCookies } from "react-cookie"
 import { AlertType, Cookie, User, UserRow } from "../../utils/types"
 import { useUpdateParticipantMutation } from "../../features/user/backendApi"
 import { useDispatch } from "react-redux"
-import { getErrorInfo } from "../../helper"
+import { getErrorInfo } from "../../utils/helper"
 import { handleShowAlert } from "../../utils/handleShowAlert"
 import Input from "../ui/Input"
 import SmartSelect from "../ui/SmartSelect"
