+++ src/components/modals/UpdateCohortModal.tsx
@@ -5,7 +5,7 @@ import { AlertType, Cookie, Stage } from "../../utils/types"
 import Input from "../ui/Input"
 import { z } from "zod"
 import { zodResolver } from "@hookform/resolvers/zod"
-import { getErrorInfo } from "../../helper"
+import { getErrorInfo } from "../../utils/helper"
 import { handleShowAlert } from "../../utils/handleShowAlert"
 import { useDispatch } from "react-redux"
 import { useCookies } from "react-cookie"
