+++ src/pages/Applicant/ApplicationForm.tsx
@@ -13,7 +13,7 @@ import {
   convertFormQuestionsToObject,
   getErrorInfo,
   getFormattedDate,
-} from "../../helper"
+} from "../../utils/helper"
 import { handleShowAlert } from "../../utils/handleShowAlert"
 import { useDispatch } from "react-redux"
 import { useEffect } from "react"
