+++ src/pages/Applicant/SavedApplication.tsx
@@ -2,7 +2,7 @@ import { useCookies } from "react-cookie"
 import { AlertType, Cookie, UserResponseQuestion } from "../../utils/types"
 import { useGetMyApplicationQuery } from "../../features/user/backendApi"
 import Loader from "../../components/ui/Loader"
-import { getErrorInfo } from "../../helper"
+import { getErrorInfo } from "../../utils/helper"
 import { handleShowAlert } from "../../utils/handleShowAlert"
 import { useDispatch } from "react-redux"
 import ApplicationFormQuestionPreview from "../../components/ui/ApplicationFormQuestionPreview"
