+++ src/pages/Cohort/Cohorts.tsx
@@ -10,7 +10,7 @@ import { Box, Modal, Typography } from "@mui/material"
 import { useState } from "react"
 import { handleShowAlert } from "../../utils/handleShowAlert"
 import { useDispatch } from "react-redux"
-import { getErrorInfo } from "../../helper"
+import { getErrorInfo } from "../../utils/helper"
 import { useCookies } from "react-cookie"
 import { customizeDataGridStyles } from "../../utils/data"
 import TableSkeleton from "../../components/skeletons/TableSkeleton"
