+++ src/pages/Form/AllForms.tsx
@@ -13,7 +13,7 @@ import Loader from "../../components/ui/Loader"
 import { Link } from "react-router-dom"
 import FormsSkeleton from "./FormsSkeleton"
 import { useCookies } from "react-cookie"
-import { getErrorInfo } from "../../helper"
+import { getErrorInfo } from "../../utils/helper"
 import { handleShowAlert } from "../../utils/handleShowAlert"
 import { useDispatch } from "react-redux"
 import { FormControl, MenuItem, Select, SelectChangeEvent } from "@mui/material"
