+++ src/pages/Overview/Applicants.tsx
@@ -16,7 +16,7 @@ import { useEffect, useState } from "react"
 import OverViewTable from "../../components/ui/OverViewTable"
 import Button from "../../components/ui/Button"
 import { useCookies } from "react-cookie"
-import { getErrorInfo } from "../../helper"
+import { getErrorInfo } from "../../utils/helper"
 import { handleShowAlert } from "../../utils/handleShowAlert"
 import { useDispatch } from "react-redux"
 import Loader from "../../components/ui/Loader"
