+++ src/pages/User/Coaches.tsx
@@ -9,7 +9,7 @@ import { useCookies } from "react-cookie"
 import { useDispatch } from "react-redux"
 import { FormControl, SelectChangeEvent, MenuItem, Select } from "@mui/material"
 import { handleShowAlert } from "../../utils/handleShowAlert"
-import { getErrorInfo } from "../../helper"
+import { getErrorInfo } from "../../utils/helper"
 import Loader from "../../components/ui/Loader"
 import { DataGrid, GridColDef } from "@mui/x-data-grid"
 import DeleteIcon from "../../assets/DeleteIcon"
