+++ src/pages/auth/Login.tsx
@@ -10,7 +10,7 @@ import InputField from "../../components/ui/InputField"
 import Loader from "../../components/ui/Loader"
 import { AlertType, ButtonSize, Cookie } from "../../utils/types"
 import { useDispatch } from "react-redux"
-import { getErrorInfo } from "../../helper"
+import { getErrorInfo } from "../../utils/helper"
 import { handleShowAlert } from "../../utils/handleShowAlert"
 import { useCookies } from "react-cookie"
 import { CredentialResponse, GoogleLogin } from "@react-oauth/google"
