+++ src/pages/auth/ResetPassword.tsx
@@ -7,7 +7,7 @@ import Button from "../../components/ui/Button"
 import { useForm } from "react-hook-form"
 import { useResetPasswordMutation } from "../../features/user/backendApi"
 import { AlertType, ButtonSize } from "../../utils/types"
-import { getErrorInfo } from "../../helper"
+import { getErrorInfo } from "../../utils/helper"
 import { handleShowAlert } from "../../utils/handleShowAlert"
 import { useDispatch } from "react-redux"
